requirements.txt
python-telegram-bot
requests
opencv-python
numpy
